working purge/kick due to lemons <3

+++ commands/purge.js
@@ -3,55 +3,52 @@ const Discord = require('discord.js')
 module.exports = {
     name: 'purge',
     description: 'bots purge command!',
-	commandOptions: [
-        {
-            type: 3,
-            name: "purge-amount",
-            description: "How many messages to purge!",
-            required: true
-        }
-    ],
+    commandOptions: [{
+        type: 4,
+        name: "purge-amount",
+        description: "How many messages to purge!",
+        required: true
+    }],
     global: false,
-    execute(interaction) {
-    const channel = client.channels.cache.get(interaction.channel_id),
-    amount = interaction.data.options[0].value;
+    async execute(interaction) {
+        const channel = await client.channels.fetch(interaction.channel_id),
+            amount = interaction.data.options[0].value;
 
-    channel.bulkDelete(amount)
+        if (isNaN(amount)) {
+            const resultembedn = new Discord.MessageEmbed()
+                .setColor("#3371FF")
+                .setTitle("Invalid.")
+                .setDescription("Invalid Arguments.")
 
-    function logResults(error, results) {
-        if (error) {
-            console.log(error);
-            message.channel.send("An error occurred.")
-        } else {
-
-            if(!results[0]) {
-                const resultembedn = new Discord.MessageEmbed()
-                    .setColor("#3371FF")
-                    .setTitle("Invalid.")
-                    .setDescription("Invalid Arguments.")
-
-                return client.api.interactions(interaction.id, interaction.token).callback.post({data: {
+            return client.api.interactions(interaction.id, interaction.token).callback.post({
+                data: {
                     type: 4,
                     data: {
-                            embeds: [resultembedn]
-                        }
-                    }
-                })
-            }
-
-            const resultembed = new Discord.MessageEmbed()
-                .setColor("#3371FF")
-                .setTitle(`Purged Successfully - ${search}`)
-                .setFooter(amount + " messages purged")
-
-            client.api.interactions(interaction.id, interaction.token).callback.post({data: {
-                type: 4,
-                data: {
-                        embeds: [resultembed]
+                        embeds: [resultembedn]
                     }
                 }
             })
         }
-    }
+
+        channel.bulkDelete(amount)
+            .then(() => {
+                const resultembed = new Discord.MessageEmbed()
+                    .setColor("#3371FF")
+                    .setTitle(`Purged Successfully`)
+                    .setFooter(amount + " messages purged")
+
+                client.api.interactions(interaction.id, interaction.token).callback.post({
+                    data: {
+                        type: 4,
+                        data: {
+                            embeds: [resultembed]
+                        }
+                    }
+                })
+            })
+            .catch(err => {
+                console.log(err);
+                channel.send("An error occurred.")
+            })
     }
 }
